requirements.txt
requests
py-trello
markdownify